+++ src/elspais/graph/GraphNode.py
@@ -27,7 +27,7 @@ class NodeKind(Enum):
     ASSERTION = "assertion"
     CODE = "code"
     TEST = "test"
-    TEST_RESULT = "result"
+    RESULT = "result"
     USER_JOURNEY = "journey"
     REMAINDER = "remainder"
     # Implements: REQ-d00126-A